pyproject.toml
[tool.poetry]
name = "margo"
version = "0.1.0"
description = ""
authors = ["starminalush <alica154323@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
tavily-python = "^0.5.0"
langgraph = "0.2.60"
langchain-community = "^0.3.8"
arxiv = "^2.1.3"
langchain-openai = "^0.2.10"
beautifulsoup4 = "^4.12.3"
chromadb = "^0.6.0"

[tool.poetry.group.linters.dependencies]
ruff = "^0.8.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
src = ["src"]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false

docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
case-sensitive = true
